# Test performance of having qthreads use Chapel's allocator