--- a/util/cron/test-perf.chapcs.playground.bash
+++ b/util/cron/test-perf.chapcs.playground.bash
@@ -9,11 +9,11 @@ source $CWD/common-perf.bash
 
 export CHPL_NIGHTLY_TEST_CONFIG_NAME="perf.chapcs.playground"
 
-# Test performance of limiting tls usage in task spawn
+# Test performance of having qthreads use chapel's allocator
 GITHUB_USER=ronawho
-GITHUB_BRANCH=avoid-tls-in-task-spawn
-SHORT_NAME=avoidTLS
-START_DATE=01/10/17
+GITHUB_BRANCH=use-chpl-alloc-for-qt
+SHORT_NAME=qtChplAlloc
+START_DATE=01/20/17
 
 git branch -D $GITHUB_USER-$GITHUB_BRANCH
 git checkout -b $GITHUB_USER-$GITHUB_BRANCH